Back_py/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.1
flask-marshmallow==0.15.0
passlib==1.7.4
python-dotenv==1.0.0
gunicorn==20.1.0
PyMySQL==1.1.0
